Add a first basic version of an AST
The current implementation of the AST takes a JSonNode that represents
the blocks of a script. The JSonNode basically holds a list of JSonNodes,
each of those representing one block. These JSonNodes are then transformed
into AST Nodes.
Currently a script only containing motion_movesteps can be read in.
For further blocks, new transformers have to be written and registered
in the dispatcher.
Currently the AST generation is not integrated in the rest of litterbox.