Implement TurnDegreesTransformer and register it at the dispatcher
This does not yet work for all turn-blocks as there are too many opcodes for the current implementation of the dispatcher, a possible fix would be to use a set of identifiers instead of a single String.
Also the input of a SingleIntInputBlock can also be a variable, we don't cover this case yet.