Handle list inputs and obscured inputs (shadows) - closes #18 and #20
SingleIntInputBlocks now work with every possible input:
Ints, variables and lists.
They also store the information about the obscured inputs in case
a variable or list obscures the int input.
Tests were added to make sure this works for every SingleIntInputBlock.
Many classes which are part of this commit didn't really change but
were affected by the renaming of the SingleIntInputBlock field "inputVariableID".