Huge refactoring - implement slots and inputs (Closes #21)
Slots and their inputs now have their own classes so that
representing them properly is possible even if there are
nested block inputs.
As things got more complex the parsing in the Ast class
was refactored and some comments were added, as well.
Parsing is fully implemented for SingleIntInputBlocks
but this also means that every SingleIntInputBlock had to be
refactored and is therefore included in this commit.
Also some of the test cases relied on the old structure and
are waiting for an update. In they meantime they live a happy
life as comments.