Refactor exception handling
This closes #67. There actually are many times where we catch exceptions
and continue parsing so this was everything I could do at this point
in terms of deleting catch blocks.

The extractProgram() method is the only method in which the
ProgramParser::parseProgram() is called (besides tests) and therefore
is the new place for catching RuntimeExceptions in order to continue analysis
if they occur.