Simplify issue finder
Nina made the great observation that this issue can be identified by simply looking at the statement list of loops and checking if the last block is a stop block. If the block was guarded by a an if-block the if-block would be the last block in the statement list of the loop, not the stop block. So any stop block that is the last is unguarded and thus this is an instance of this issue.