parser: pass state within parser instead of using a shared mutable one
In the previous implementation it was not possible that multiple threads
parse different programs at the same time due to the shared/static
mutable state in the ProgramParser and ActorDefinitionParser.
This is solved by creating a state only once when starting to parse a
program and passing this one around to the parsers for the different
elements. Therefore, the state is contained within the parser of a single
program and multiple programs can be parsed at the same time.