refactor analyzer inheritance structure
NOTE:
This is a breaking change for users of the LitterBox via the API.
CLI-users are not affected by these changes.

The previous `Analyzer` interface mixed the ability to analyse already parsed
programs, programs stored in files, and to-be-downloaded projects based on their
project id. Especially when using such analyzers via the API directly rather
than via the CLI, one had to pass in `null`-values as input/output parameters.
The previous approach also did not really separate concerns (analysis logic, vs.
rendering the results into some structured format for file writing).

The base analyzer is now defined by an interface that defines the analysis step
as something that takes in an already parsed program and returns some result
object. This is extended by a `FileAnalyzer` that reads the project from a file
and produces output to a file (or the console) again. A further extension is the
`ProjectIdAnalyzer` that only needs one or multiple project IDs, downloads the
corresponding Scratch project(s) to a file, and then acts like a `FileAnalyzer`
on them.

Additionally, the `FileAnalyzer` no longer takes in the `input` file as part of
the constructor, but as part of the `analyze` method. This makes the same
analyzer reusable for multiple invocations.