fix(parser): make more resilient against filenames without extension
The parser would previously crash when given project file names without
file extensions.
To fix this, we now make no assumptions about the existence of file
extensions any more. In case there is no file extension, the program
name will be the full file name. The same logic of using the file name
with removed extension therefore remains regardless of the existence of
an extension.