fix: parse all custom block %n parameters as %s
When creating new custom procedures in the Scratch UI, block inputs of
type number/text are crated as '%s'. The '%n' probably exists for
historical reasons when converting from Scratch 2 programs. The Scratch
VM interprets them as '%s' anyway.

To make the call/procedure pair lookup easier and more consistent, we
now parse all '%n' parameters as '%s'. The actual parameter type was
already parsed as 'StringType', but the name of the method/call was not
adapted to match this behaviour.