scratchblocks(parser): linearise binary expressions
This improves the parsing speed for both binary and number expressions.
In both (binary and numeric) cases, the expressions share a common
prefix. Therefore, the parser already tries to parse the (potentially
deeply nested) left side of the operator before noticing a mismatch for
the operator itself and having to backtrack. Consolidating the rules
into one and moving the choice of the operator inside the rule fixes
this problem.