make llm analysers independent of ScratchLlm
Adds properties to choose the LlmApi and PromptBuilder via system
properties. This allows the construction of the analysers without
explicitly specifying them. The analysers still provide the option to
manually instantiate them with specific LlmApi and PromptBuilder
instances. This (a) makes them easier to test, and (b) allows for
more flexibility when using LitterBox as a library and using the
internal Java API directly.