hash issues based on the ID
The ID should be globally unique by construction. Therefore, we can use
it to hash the issue, eg when constructing sets of issues and also when
checking for equality.
The IssueBuilder can now also be used to clone existing issues while
keeping their IDs the same. This is for example useful when the original
issue is used as a base issue but the hint text is later extended. In
this case it should be possible to explicitly retain the issue ID since
it is in its core still the same issue as before.