fix property loading in JSONReportGenerator
Statically setting these options should work for most usecases. However,
during testing, we set different values for these properties. Therefore,
this breaks when statically loading such values *once* when initialising
the class for the first time. Now, we instead load the values every time
a new instance of the `JSONReportGenerator` is created. This is okay for
normal use performancewise, since when running LitterBox as CLI program,
the report generator is only ever instantiated once to write the report.